docs: add documentation
