+++ lib/src/header_box/header_box.dart
@@ -1,6 +1,7 @@
 export 'sbb_header_box.dart';
 export 'sbb_header_box_flap.dart';
 export 'sbb_sliver_header_box.dart';
+export 'sbb_sliver_header_box_config.dart';
 export 'sbb_sliver_header_box_spacer.dart';
 export 'sliver/sbb_cascade_column.dart';
 export 'theme/sbb_header_box_flap_style.dart';
